… .github/workflows/build.yml
… CMakeLists.txt
… examples/simple/simple.cpp
… ggml-backend.h
… ggml-cuda-quant.h
… ggml-cuda.cu
… ggml.c
… ggml.h
… llama.cpp