This PR adds a common interface to the compute backends.

### Breaking changes
- `ggml_context` allocates memory from a `ggml_buffer` that contains a buffer in device memory for the tensor data, and a buffer in system memory for the tensor structs (and in the future also other data such as the graphs)
- The `data` member of `ggml_tensor` is a backend-specific pointer that should not be accessed directly. To access the data, `ggml_backend_set_tensor` and `ggml_backend_get_tensor` must be used instead. Functions such as `ggml_new_f32` and `ggml_set_f32` can also be used as before.
  - Technically, if you are using the CPU backend, you can still access the `data` member directly, but you shouldn't do that if you want to support other backends
  - I will probably change the name to something else to prevent current code from compiling without changes
- Added a small `params` buffer to `ggml_tensor` for the op parameters that currently are stored in a tensor. For example, for `ggml_rope` this buffer is used to store the values `n_past, n_dims, mode, n_ctx`. The goal is to make these parameters easily accessible from the CPU, and reduce the overhead of creating a new tensor for them.

Brief example:
```C
// initialize a backend
struct ggml_backend backend = ggml_backend_cpu_init();
// for CUDA:
// struct ggml_backend backend = ggml_backend_cuda_init();

// create a buffer
size_t ctx_size = 4096; // buffer size for the tensor data
size_t num_tensors = 10; // maximum number of tensors that can be allocated
struct ggml_buffer buf = ggml_backend_alloc_buffer(&backend, ctx_size, num_tensors);

// create a context using the buffer
struct ggml_init_params ggml_params = ggml_init_params_default();
ggml_params.buffer = &buf;
ggml_context * ctx = ggml_init(ggml_params);

// use the context to create a computation graph
struct ggml_tensor * x = ggml_new_f32(ctx, 2.0f);
struct ggml_tensor * a  = ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 1);
struct ggml_tensor * b  = ggml_new_tensor_1d(ctx, GGML_TYPE_F32, 1);
struct ggml_tensor * x2 = ggml_mul(ctx, x, x);
struct ggml_tensor * f  = ggml_add(ctx, ggml_mul(ctx, a, x2), b);
struct ggml_cgraph gf = ggml_build_forward(f);

// set the value of the input tensors
float a_val = 3.0f;
ggml_backend_set_tensor(a, &a_val, 0, sizeof(float));
// alternatively:
float b_val = 5.0f;
ggml_set_f32(b, b_val);

// run the computation
ggml_backend_graph_compute(&backend, &gf);

// get the result
float result;
ggml_backend_get_tensor(f, &result, 0, sizeof(float));
// alternatively:
// result = ggml_get_f32_1d(f, 0);
```

### Backend implementation
Backends should implement the functions defined in the `ggml_backend_interface` struct. Currently there are implementations for the CPU and CUDA backends.

### Computation using multiple backends
It is still possible to offload some parts of the graph to the GPU while keeping others on the CPU. This is done using `ggml_graph_splits`. See the llama.cpp code for an example, will update this later with more details.

### Notes/limitations
- Only the CPU and CUDA backends are currently supported.
- Only the bare minimum necessary to run llama is currently implemented, don't look too much into the details of the code for now
- When partially offloading a model to the GPU, internally this is handled by splitting the graphs into multiple parts, and running each of them in sequence in a different backend. Because the KV memory is either on the CPU or on the GPU, this means that there are at least as many graph executions as there are layers. The CPU backend creates and destroys the threads with every graph launch, so this can have a non-negligible overhead. Eventually this will be improved by using a thread pool.
- On the plus side, the CPU threads are no longer spinning while the GPU backend is running
- In the long term, the goal is to support automatic fallback to the CPU when an operation is not implemented in a backend, but that's going to take a while. This means that backends like OpenCL that mostly only implement matrix multiplication cannot be used for now. I am expecting that this backend will be replaced by a Vulkan backend that implements full GPU offloading, and won't be necessary anymore.
- For prompt processing, it is usually worth to upload the weights every time even when using partial offloading, but that's currently not supported. So prompt processing will use the same CPU/GPU layer split as generation. Eventually I would like to support this again.