@slaren 2023-07-15
Updated the description with more details.
@ggerganov @JohannesGaessler @0cc4m @niansa @evanmiller and anybody else who may be interested in working on a backend, please take a look at the backend interface in `ggml-backend.h` and let me know that would work for your case. There are implementations for the CPU backend in `ggml-backend.c`, and for the CUDA backend in `ggml-cuda.cu`.