@JohannesGaessler 2023-07-15
>The CPU backend creates and destroys the threads with every graph launch, so this can have a non-negligible overhead. Eventually this will be improved by using a thread pool.

Do you mean to say that *you* will implement it? I'm asking because I was thinking of doing this myself and would like to avoid duplicate work.