@0cc4m 2023-07-15
Please correct me if I'm wrong, but it looks like the `ggml-backend.h` is only for backends that can run an entire (part of a) graph. While this may be the end goal for most backends, they are built up through many intermediate steps that don't fit into this framework. OpenCL cannot offload a full graph, and neither can my Vulkan approach (for now).

Do you have a plan on how to accomodate backends in development and "partial backends", so to say?