@slaren 2023-07-15
@0cc4m I covered it briefly in the description:
>In the long term, the goal is to support automatic fallback to the CPU when an operation is not implemented in a backend, but that's going to take a while. This means that backends like OpenCL that mostly only implement matrix multiplication cannot be used for now. I am expecting that this backend will be replaced by a Vulkan backend that implements full GPU offloading, and won't be necessary anymore.

For now, this is not supported, but I would like to support it eventually.