@0cc4m 2023-07-15
It *may* be replaced *eventually* by a Vulkan backend, or it may stay alongside it. But even the Vulkan backend will stay partial for quite a while. Is your plan to implement this fallback before this PR is ready to merge?