@slaren 2023-07-15
I will think about it. My plan is to do it later, the scope of this PR is already quite large as it is, but maybe I can support that case without too many changes.