@0cc4m 2023-07-15
Removing other backends in the name of a cuda refactor is not acceptable. I support the goal of this PR, but it has to be less disruptive.