@niansa 2023-07-15
Mmmh the idea really makes a lot of sense and I like it, but I feel like different implementations may have different needs that aren't covered by this and are still going to fall back to `#ifdef`ing in `llama.cpp` or `ggml.c`...