@slaren 2023-07-15
@niansa I would like to avoid that, which is why I am asking for your feedback. If there is something that isn't covered by this API, please let me know.