@JohannesGaessler 2023-07-15
I think as of right now the OpenCL implementation still has all data except for the weights in RAM. In other words, the results get copied to RAM after each operation. Would it be possible to just forward the non GPU accelerated tensors to the CPU backend for the OpenCL backend?