@slaren 2023-07-15
Fallback to CPU would work exactly like that, the hard part is making it in a way that doesn't litter the user code with backend-specific details, and also keeps the implementation of the backends as simple as possible.
Really, this could already be done with this implementation is you just create a new graph split for each matrix multiplication. What the splits interface does is just to copy the data between backends as required. But obviously that would make the graph building code unreadable, so it needs to be done in a more automated way. It can be done, but the scoped PR needs to be kept reasonable or it will never be completed.