@henk717 2023-07-15
OpenCL is currently relied on by AMD and Intel users for speed increases using their dedicated GPU. Removing it would block off entire vendors on Windows and should not be considered acceptable. Especially users with slower CPU's would be stuck on older versions of the project.

We have a lot of AMD users currently relying on Llamacpp to use their GPU for AI as an easy solution.