@JohannesGaessler 2023-07-15
What would prevent us from keeping the status quo for OpenCL? From what I can tell the entry points for OpenCL kernels is still in the corresponding kernels in `ggml.c`. So couldn't you run a CPU graph with some OpenCL tensors in that graph for the offloaded weights?