@slaren 2023-07-15
It could work if we keep two versions of the model loading code in llama.cpp, one for the new interface and another for OpenCL. It's something I would prefer to avoid, ultimately the goal is to simplify the code and remove as many of the special cases as possible, but as a temporary solution it may be better than removing the OpenCL backend.