@slaren 2023-07-15
It would be interesting to see performance on native Linux, and also native Windows. For me **under WSL** with a RTX 3080, this is significantly faster than master. I think this mainly because it requires less synchronization, so kernel launches can be queued and this helps hiding the latency of launching a kernel. I expect that the effect will be minimal if any under native Linux, though.

```
PR:
-ngl 99: llama_print_timings:        eval time =  1163.00 ms /   127 runs   (    9.16 ms per token,   109.20 tokens per second)
-ngl 30: llama_print_timings:        eval time =  2083.09 ms /   127 runs   (   16.40 ms per token,    60.97 tokens per second)
-ngl 20: llama_print_timings:        eval time =  6841.63 ms /   127 runs   (   53.87 ms per token,    18.56 tokens per second)
-ngl 10: llama_print_timings:        eval time = 13264.34 ms /   127 runs   (  104.44 ms per token,     9.57 tokens per second)


master:
-ngl 99: llama_print_timings:        eval time =  3140.24 ms /   127 runs   (   24.73 ms per token,    40.44 tokens per second)
-ngl 30: llama_print_timings:        eval time =  4398.86 ms /   127 runs   (   34.64 ms per token,    28.87 tokens per second)
-ngl 20: llama_print_timings:        eval time =  8466.26 ms /   127 runs   (   66.66 ms per token,    15.00 tokens per second)
-ngl 10: llama_print_timings:        eval time = 13212.88 ms /   127 runs   (  104.04 ms per token,     9.61 tokens per second)
```