@JohannesGaessler 2023-07-15
>I think this mainly is because it requires less synchronization, so kernel launches can be queued and this helps hiding the latency of launching a kernel.

With all layers offloaded there should only be 2 calls to `cudaDeviceSynchronize` though: one at the end and another one when the results of `embeddings` get copied to RAM. If anything I think the difference comes from concurrent kernel execution.