@slaren 2023-07-15
The concurrent kernel execution is not enabled currently:
https://github.com/ggerganov/llama.cpp/blob/83595ecbd660becf2cedc2a7bad52db7085e1373/ggml-cuda.cu#L1-L2
This is because currently it creates way too many streams and events, and it may actually harm performance. But it will help once that is fixed.