@JohannesGaessler 2023-07-15
Regardless of the reasons, simple testing seems to confirm that the refactor provides a speedup:

| GPU      | Model    | Test  | t/s master | t/s PR | Speedup |
|----------|----------|-------|------------|--------|---------|
| RTX 3090 | 7b q4_0  | tg128 |     121.64 | 132.45 |    1.09 |
| RTX 3090 | 13b q4_0 | tg128 |      69.59 |  74.07 |    1.06 |